Use a block display title rather than the default core block name in the drop down label, similar to the way the block list view does it. Add a maximum truncate value Updated tests
