+++ packages/block-editor/src/components/block-title/index.js
@@ -1,23 +1,10 @@
 /**
- * External dependencies
- */
-import { truncate } from 'lodash';
-
-/**
- * WordPress dependencies
- */
-import { useSelect } from '@wordpress/data';
-import {
-	getBlockType,
-	__experimentalGetBlockLabel as getBlockLabel,
-	isReusableBlock,
-} from '@wordpress/blocks';
-
-/**
  * Internal dependencies
  */
-import useBlockDisplayInformation from '../use-block-display-information';
-import { store as blockEditorStore } from '../../store';
+
+import useBlockDisplayTitle, {
+	MAXIMUM_TITLE_LENGTH,
+} from './use-block-display-title';
 
 /**
  * Renders the block's configured title as a string, or empty if the title
@@ -31,50 +18,13 @@ import { store as blockEditorStore } from '../../store';
  *
  * @param {Object} props
  * @param {string} props.clientId Client ID of block.
+ * @param {number} props.maximumLength The maximum length that the block title string may be before truncated.
  *
- * @return {?string} Block title.
+ * @return {JSX.Element} Block title.
  */
-export default function BlockTitle( { clientId } ) {
-	const { attributes, name, reusableBlockTitle } = useSelect(
-		( select ) => {
-			if ( ! clientId ) {
-				return {};
-			}
-			const {
-				getBlockName,
-				getBlockAttributes,
-				__experimentalGetReusableBlockTitle,
-			} = select( blockEditorStore );
-			const blockName = getBlockName( clientId );
-			if ( ! blockName ) {
-				return {};
-			}
-			const isReusable = isReusableBlock( getBlockType( blockName ) );
-			return {
-				attributes: getBlockAttributes( clientId ),
-				name: blockName,
-				reusableBlockTitle:
-					isReusable &&
-					__experimentalGetReusableBlockTitle(
-						getBlockAttributes( clientId ).ref
-					),
-			};
-		},
-		[ clientId ]
-	);
-
-	const blockInformation = useBlockDisplayInformation( clientId );
-	if ( ! name || ! blockInformation ) return null;
-	const blockType = getBlockType( name );
-	const blockLabel = blockType
-		? getBlockLabel( blockType, attributes )
-		: null;
-	const label = reusableBlockTitle || blockLabel;
-	// Label will fallback to the title if no label is defined for the current
-	// label context. If the label is defined we prioritize it over possible
-	// possible block variation title match.
-	if ( label && label !== blockType.title ) {
-		return truncate( label, { length: 35 } );
-	}
-	return blockInformation.title;
+export default function BlockTitle( {
+	clientId,
+	maximumLength = MAXIMUM_TITLE_LENGTH,
+} ) {
+	return useBlockDisplayTitle( clientId, maximumLength );
 }
